ZIO as specs2 Fragment for tests
